C:/Program Files/Git/users/{id} returns json

+++ Domain_Objects/src/main/java/potatoes/project/repository/UserRepository.java
@@ -6,6 +6,6 @@ import potatoes.project.domain_objects.User;
 
 public interface UserRepository extends JpaRepository<User, Long>{
 	User findByName(String name);
-    User findByUserID(int userID);
+    User findByUserID(long userID);
     User findByEmail(String email);
 }
